Moved async from refactoring
