+++ src/core/async/README.md
@@ -1,0 +1,3 @@
+# core/async
+
+This module extends the [Async](https://v4fire.github.io/Core/modules/src_core_async_index.html) class from and adds a bunch of methods for browser specific async tasks.
